Carry fractions in the stats so real-time decay actually happens
The full audit found the game frozen at real time, and the freeze was total:
the kernel watchdog invokes the capsule every 5 seconds, a 5-second span decays
fullness by 0.011 points, drop_stat rounded that back to the same integer — and
last_seen_ms still advanced, consuming the elapsed time. As long as the daemon
was running, no stat ever moved at scale 1.0. Compressed-time demos (5 s = 4+
pet-hours) masked it completely, which is why every live verification so far
passed while the actual game did not work.

Stats are now f64 end to end: drop/raise are pure clamps with no rounding, so
sub-point decay accumulates honestly across any polling rate. The outside world
still sees integers — PetView, bars, messages and the prompt section round at
the boundary — and stats_unchanged compares rounded values so a 0.2-point drift
does not count as a player-visible change. Saved pets migrate for free: serde
parses the stored integers straight into f64.

New regression test drives 720 five-second advances against one one-hour
advance and demands they agree — the exact shape of the freeze. Verified live:
fullness fell 99.963 to 99.860 over 45 real seconds, the first observed
real-time decay in the capsules

+++ src/model.rs
@@ -9,23 +9,26 @@ pub const STATE_VERSION: u32 = 1;
 pub const MAX_ALERTS: usize = 20;
 pub const MAX_NAME: usize = 32;
 /// Stats a freshly adopted pet starts with.
-pub const START_STAT: u8 = 80;
+pub const START_STAT: f64 = 80.0;
 
 fn default_version() -> u32 {
     STATE_VERSION
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+// Stats are f64 so sub-point decay accumulates honestly: the 5 s watchdog tick
+// spans decay fractions of a point, and integer stats rounded them all away.
+// No `Eq` — f64 is only `PartialEq`.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Pet {
     #[serde(default = "default_version")]
     pub version: u32,
     pub name: String,
     pub born_at_ms: u64,
     pub last_seen_ms: u64,
-    pub fullness: u8,
-    pub happiness: u8,
-    pub energy: u8,
-    pub cleanliness: u8,
+    pub fullness: f64,
+    pub happiness: f64,
+    pub energy: f64,
+    pub cleanliness: f64,
     #[serde(default)]
     pub sleeping: bool,
     #[serde(default)]
