+++ src/moment.rs
@@ -124,10 +124,10 @@ impl Active {
 
 /// Which moments a pet's condition currently allows.
 #[must_use]
-pub fn eligible(fullness: u8, happiness: u8, energy: u8, cleanliness: u8) -> Vec<u16> {
-    let thriving = fullness >= 70 && happiness >= 70 && energy >= 70 && cleanliness >= 70;
-    let neglected = happiness < 30;
-    let energetic = energy >= 60;
+pub fn eligible(fullness: f64, happiness: f64, energy: f64, cleanliness: f64) -> Vec<u16> {
+    let thriving = fullness >= 70.0 && happiness >= 70.0 && energy >= 70.0 && cleanliness >= 70.0;
+    let neglected = happiness < 30.0;
+    let energetic = energy >= 60.0;
 
     MOMENTS
         .iter()
@@ -223,7 +223,7 @@ mod tests {
 
     #[test]
     fn gates_keep_sad_moments_away_from_a_thriving_pet() {
-        let happy = eligible(90, 90, 90, 90);
+        let happy = eligible(90.0, 90.0, 90.0, 90.0);
         let keys: Vec<_> = happy
             .iter()
             .filter_map(|i| MOMENTS.get(*i as usize))
@@ -236,7 +236,7 @@ mod tests {
 
     #[test]
     fn a_neglected_pet_can_have_the_sad_moments() {
-        let sad = eligible(10, 10, 10, 10);
+        let sad = eligible(10.0, 10.0, 10.0, 10.0);
         let keys: Vec<_> = sad
             .iter()
             .filter_map(|i| MOMENTS.get(*i as usize))
@@ -250,7 +250,12 @@ mod tests {
     #[test]
     fn every_moment_is_reachable_from_some_condition() {
         let mut seen: Vec<&str> = Vec::new();
-        for (f, h, e, c) in [(90, 90, 90, 90), (10, 10, 10, 10), (50, 50, 20, 50), (50, 50, 90, 50)] {
+        for (f, h, e, c) in [
+            (90.0, 90.0, 90.0, 90.0),
+            (10.0, 10.0, 10.0, 10.0),
+            (50.0, 50.0, 20.0, 50.0),
+            (50.0, 50.0, 90.0, 50.0),
+        ] {
             for i in eligible(f, h, e, c) {
                 if let Some(m) = MOMENTS.get(i as usize) {
                     seen.push(m.key);
@@ -264,7 +269,7 @@ mod tests {
 
     #[test]
     fn pick_always_returns_something_eligible() {
-        let set = eligible(50, 50, 50, 50);
+        let set = eligible(50.0, 50.0, 50.0, 50.0);
         assert!(!set.is_empty());
         for seed in [0u32, 1, 7, 99, 12345, u32::MAX] {
             let got = pick(&set, seed).expect("must pick");
@@ -279,7 +284,7 @@ mod tests {
 
     #[test]
     fn weights_actually_bias_the_outcome() {
-        let set = eligible(50, 50, 50, 50);
+        let set = eligible(50.0, 50.0, 50.0, 50.0);
         let mut sunbeam = 0;
         let mut philosopher = 0;
         for seed in 0..4000u32 {
